requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
fastapi==0.111.0
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
Jinja2==3.1.4
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
robotframework==7.0.1
robotframework-datadriver==1.11.2
robotframework-requests==0.9.7
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
websockets==12.0
robotframework-pabot==2.18.0
sqlalchemy==2.0.31